test: move the --inspect-brk breakpoint regression tests to a dedicated file
Put the #32591 regression tests in test/cli/inspect/inspect-brk-breakpoint.test.ts,
alongside the other topic-specific inspector test files (HTTPServerAgent,
BunFrontendDevServer, test-reporter), instead of in inspect.test.ts.
